Chrisitan avec un bug sur le readlong de client mais marche avec telnet

+++ log736-labo1/Cristian_API/interfaces/IClient.java
@@ -6,12 +6,12 @@ import java.net.Socket;
 
 public interface IClient {
 
-    public void requestTime(int serverPort, long currentTime, int numberOfTries) throws java.io.IOException;
+    public  void requestTime(int serverPort, long currentTime, int numberOfTries) throws java.io.IOException;
 
-    public long getTime();
+    public  long getTime();
 
     public void setTime(long newTime);
 
-    public int getAccuracy();
+    public long getAccuracy();
 
 }
